requirements_stable.txt
matplotlib==2.0.0
PyTorch
litus==0.1
future
